# optimize is_token function by using a bit table lookup